Turn on polymode for certain file extensions
Also change the name of the generated files.  By default the files have
the characters '[' and ']' in them, e.g., my_file[woven].tex.  This
causes problems when zsh is used to run the weave/export command because
these characters have special meaning.  So, they should be escaped or
quoted.  I couldn't immediately figure out how to do this in the elsip
code to submit a pull request.