Create lite versions of Emacs startup files
Emacs take a long time to start up with my initialization files.  This
is not a problem for desktop usage, because the daemon only needs to
start once.

When the daemon is not running and I want to quickly edit a file, these
lite versions allow Emacs to start quickly.  I use them with the
el (Emacs lite) shell alias.

alias el="emacs -nw -q -l ~/.emacs.d/init-lite.el"