Add (commented out) Stump commands to make Firefox default browser
Conkeror is dying because libxul is dead. I am trying to continue using it with
Waterfox, but that project is not keeping up.  For now, stick with Conkeror,
because Firefox is too painful, but the end is neigh for Conkeror.  Probably the
last painful path forward is to use one of Qutebrowser, webmacs, or Next, but
they still have a long way to go to catch up to how great Conkeror was.