.zshrc: Re-add ta alias, remove ta(), and restore prexec()
Remove the flawed ta() function.  It was flawed, because after re-attaching
commands are not sent to the attached shell.  I cannot find a robust way to do
this automatically, once when re-attaching.  Solutions with 'tmux send-keys' or
'paste-buffer' are fragile.  So, go back to running something with every shell
command in preexec(), even though it feels invasive.